Fix cc warnings; print src version

+++ src/soundcard/sound.cxx
@@ -660,7 +660,7 @@ int cSoundPA::Open(int mode, int freq)
         cerr << "cSoundPA::Open(" << mode << ", " << freq << "); req_sample_rate="
              << req_sample_rate << " sample_frequency=" << sample_frequency << endl;
 
-        int old_sample_rate = req_sample_rate;
+        int old_sample_rate = (int)req_sample_rate;
         req_sample_rate = sample_frequency = freq;
 
         // Try to keep the stream open if we are using jack, or if we
